Add new vnc script (hyakvnc.py) for Klone
This commit introduces a python3-based vnc start/end script that is much
more robust than the previous version.

This version attempts to address port-conflict issues that were present
in the bash version.

This version also prompts the user to set password instead of requiring
the user to run a command inside the xfce container manually.

Notes:

1. Change constants (singularity binary location, xfce container path,
etc.) if necessary.

2. Run setup.sh once to get all python dependencies using the xfce
container.

3. hyakvnc.py has many options; Run "./hyakvnc.py -h" to see options.