hyakvnc.py: Update VNC display number parser
This commit fixes an issue with the display parser for users with
multiple VNC session instances.

Issue is caused by matching against the first display number in the list
(shown by running `vncserver -list`). This issue can cause the parser to
report an incorrect display number for the session being created. And
with an incorrect display number, an incorrect port forward will be
created.

The issue is fixed by parsing directly from the first command since the
association between session and display number is more clearer.