[TEMP] hyakvnc.py: reserve exclusive nodes for vnc session
This temporary commit forces each job allocation to allocate a node that
has not been allocated. This will be temporary until a method for
associating a job id with its vnc session can be determined. Finding a
method may prove to be difficult (unless there is a reliable method for
running commands for a specific job) since `ssh <node> <command>` does
not make the distinction for which job id it will execute commands on.
Hopefully, I am misunderstanding what jobs are and that there is a
method available.