hyakvnc.py: Add --repair flag to repair missing/broken port forwards
This new flag will attempt to re-create port forwards for vnc jobs that
are missing port forwards. This assumes an active vnc process for each
affected job. If there is no vnc process, then it is recommended to kill
the job and start from scratch.

Useful when login node is forced to reboot which kills
LoginNode<->SubNode port forwardings.